Bugfix: reboot needed for dock audio checkbox 1/2
After the patch
http://review.cyanogenmod.com/#/c/15540/

The Dock audio is only observed if the checkbox is checked onBootComplete

This patch adds an extra Event for when the dock setting is changed,
this is picked up by the WiredObserver and either stops or starts listening to the dock.

Note that stopObserving() stops all observations so the other endpoints need re-observing.

Change-Id: Ifcb3e36f116688d0ccb7be15f1e7aa3d24e2981c